Working arbitrary node dockerized cosmos chain
This commit generates a working arbitrary node cosmos chain after
building the required binaries from scratch. Very nice for testing.
It still needs a lot of convience features but it's working well enough
to be good for now.

Next up is the addition of the Etherum chain and deployment of the
peggy contracts.