Add REST endpoints for validator-daemon
The cli endpoints do the signing for you, this makes sense in the
context of making them ergonomic to use. But for actual production
use we're going to have the validator daemon call REST endpoints
in order to manage Peggy operation.

In that case we probably want to kee the ETH private key within
the validator daemon and do the signing there. So these endpoints
verify the sigs and public messages as required.