Add test runner
Test runner is a rust library that will be handling testing of the
Cosmos code and various other tasks.

Usually we would make such a test script in Python, but that's not
really practical here beucase we want to do things like test CosmosSDK
transaction submission and the on-chain state machine in general.

This means we need infrastructure to call the rest endpoints, validate
sigs, and make sigs. These are all things we need for the validator
daemon as well.

If we combine the tooling we can have the test runner serve as a testing
platform for the validator daemon and perform the testsing tasks we need
all without adding another set of signing and verifying code to our code
burden. I think this is the best decision even if it leads to writing
'Pythonesque' Rust.