No signatures in Peggy constructor
this modifies the Peggy contract to not require signatures in order
to deploy a new contract.

This comes with upsides and downsides. On one hand there is less
vailidation before deployment. On the other hand we get to reduce
the state machine complexity on the blockchain side. Since we don't
need to be making validator set updates before the contract is even
known.

In the end I think this is the right trade off. But I've placed it
into this nicely revertable commit just in case.